Abstract Tree Component to seperate file
